Reorganize core objects
Now in their own folder, and each object separated into its own script.
