+++ nept/core/__init__.py
@@ -1,0 +1,6 @@
+from .analogsignal import AnalogSignal
+from .epoch import Epoch
+from .localfieldpotential import LocalFieldPotential
+from .neurons import Neurons
+from .position import Position
+from .spiketrain import SpikeTrain
