+++ nept/core/tests/test_analogsignal.py
