+++ nept/core/tests/test_spiketrain.py
